[VCDA-1894] Refactor cse.py and separate out different command groups (#827)
* initial changes - created files for command groups


* Remove unnecessary imports


* rename modules to x_commands.py


* Fixed copyright year in new files
